Ignore custom app space management activities
Allowing apps to override the "Clear storage" function makes users
powerless when it comes to clearing the data of a rogue app, short of
reinstalling the entire app. I've never seen an app that uses this in a
good way, so just remove it to allow clearing storage of any app.

The most prominent example of this is the Google Search (Velvet) app,
which does not allow the user to clear all data and leaves up to 100 MB
of wasted space even after clearing storage in its custom space manager.
It's also impossible to clear its data after disabling the app, which
forces users to deal with wasted space for no reason.

Change-Id: Id8a238b0838c99e63086ac8d5a6c1d8a7609494d
Signed-off-by: ryzenforce990 <ryzenforce990@gmail.com>